Update jquery.webui-popover.js
pass the event when hiding so that we are able to know if the hiding process started from the element itself or from the body or the close icon. In case the hiding started from the element itself, we can immediately trigger its underlying function. Otherwise we do not trigger it.